requirements_test.txt
-r requirements_test_pre_commit.txt

async-upnp-client==0.46.0
aiohomematic==2025.12.9
aiohomematic-test-support==2025.12.9
isal==1.8.0
mypy==1.18.2
pre-commit==4.5.0
pur==7.3.3
pylint==4.0.4
pylint_strict_informational==0.1
pytest-homeassistant-custom-component==0.13.299
